Fix nondeterminism on Activity/Suborchestration failures
When we wrap exceptions from the C# layer (i.e. Activity or
SubOrchestration failures), we cannot fail the function execution. If we
do, the C# layer will not have the chance to replay all of the actions
performed by the JS SDK, meaning we will result in a non-determinism
error.

To avoid this, we mark the JS execution as successful, and put the
burden on the C# shim for out-of-proc languages to surface these
failures.