Fix nondeterminism errors when Orchestration throws Error (#145)
When we throw errors in the JS layer, we need some way to embed the orchestration replay state in the errors so that the C# extension can replay behavior. This is necessary so that the Durable Task does not throw an non-determinism exception.

This also requires a fix on the C# extension to be able to read this embedded data from the RPC exception. Here is the PR for the fix in Durable 1.x: Azure/azure-functions-durable-extension#1171.

Resolve #125